+++ tests/testthat/test_matching.R
@@ -11,7 +11,8 @@ test_that("closest throws errors", {
     expect_error(closest(1:3, 1:3, nomatch = TRUE), "be a 'numeric'")
     expect_error(closest(1, 1, nomatch = 1:2),
                  "'nomatch' has to be a 'numeric' of length one")
-    expect_error(closest(1:3, c(1, NA)), "not  contain NA")
+    expect_error(closest(1:3, c(1, NA)), "not contain NA")
+    expect_error(closest(1:3, 1:3, duplicates = "foo"), "has to be .*keep.*,")
 })
 
 test_that("closest basically works", {
@@ -72,6 +73,19 @@ test_that("closest, duplicates", {
     expect_equal(closest(1.6, 1:2, tolerance = 0.5, duplicates = "keep"), 2)
     expect_equal(closest(1.6, 1:2, tolerance = 0.5, duplicates = "closest"), 2)
     expect_equal(closest(1.5, 1:2, tolerance = 0.5, duplicates = "remove"), 1)
+
+    # equal distances, see
+    # https://github.com/rformassspectrometry/MsCoreUtils/issues/65
+    x <- as.numeric(c(1, 3, 5, 6, 8))
+    y <- as.numeric(c(3, 4, 5, 7))
+    expect_equal(closest(x, y, tolerance = 1, duplicates = "closest"),
+                 c(NA, 1, 3, 4, NA))
+    # multiple match, with better match following, see
+    # https://github.com/rformassspectrometry/MsCoreUtils/issues/66
+    x <- c(1, 1.5, 2, 2.1, 5, 6, 7)
+    y <- c(4.6, 4.7, 4.8, 4.9, 5, 6, 7, 8)
+    expect_equal(closest(x, y, tolerance = 3, duplicates = "closest"),
+                 c(NA, NA, NA, 1, 5, 6, 7))
 })
 
 test_that("common", {
@@ -87,8 +101,8 @@ test_that("common", {
 })
 
 test_that("join", {
-    x <- c(1, 2, 3, 6)
-    y <- c(3, 4, 5, 6, 7)
+    x <- as.numeric(c(1, 2, 3, 6))
+    y <- as.numeric(c(3, 4, 5, 6, 7))
 
     expect_equal(join(x, y, type = "outer"),
                  list(x = c(1:3, NA, NA, 4, NA), y = c(NA, NA, 1:5)))
@@ -146,8 +160,123 @@ test_that("join", {
                  list(x = integer(), y = integer()))
 
     ## no match at all
-    x <- c(1, 2, 3, 6)
-    y <- c(4, 5, 7)
+    x <- as.numeric(c(1, 2, 3, 6))
+    y <- as.numeric(c(4, 5, 7))
     expect_equal(join(x, y, type = "outer"),
                  list(x = c(1:3, NA, NA, 4, NA), y = c(NA, NA, NA, 1:2, NA, 3)))
+})
+
+test_that("outer join works", {
+    x <- as.numeric(c(1, 2, 3, 6))
+    y <- as.numeric(c(3, 4, 5, 6, 7))
+
+    expect_error(join(1:3, 4:1, 0, 0, type = "outer"), "sorted")
+    expect_error(join(4:1, 1:4, 0, 0, type = "outer"), "sorted")
+    expect_error(join(c(1, 2, NA, 3, type = "outer"), 1:4, 0, 0), "sorted")
+
+    expect_equal(join(x, y, 0, 0, type = "outer"),
+                 list(x = c(1, 2, 3, NA, NA, 4, NA),
+                      y = c(NA, NA, 1, 2, 3, 4, 5)))
+    expect_equal(join(x, y, 10, 0, type = "outer"),
+                 list(x = c(1, 2, 3, NA, NA, 4, NA),
+                      y = c(NA, NA, 1, 2, 3, 4, 5)))
+    expect_equal(join(y, x, 10, 0, type = "outer"),
+                 list(x = c(NA, NA, 1, 2, 3, 4, 5),
+                      y = c(1, 2, 3, NA, NA, 4, NA)))
+
+    x <- c(1, 1.5, 2, 2.1, 5, 6, 7)
+    y <- c(4.6, 4.7, 4.8, 4.9, 5, 6, 7, 8)
+    expect_equal(join(x, y, 0, 0, type = "outer"),
+                 list(x = c(1:4, NA, NA, NA, NA, 5:7, NA),
+                      y = c(NA, NA, NA, NA, 1:8)))
+    expect_equal(join(y, x, 0, 0, type = "outer"),
+                 list(x = c(NA, NA, NA, NA, 1:8),
+                      y = c(1:4, NA, NA, NA, NA, 5:7, NA)))
+    ## Issue #66: outer join with tolerance 3: expect to have more matches!
+    tol_3 <- list(x = c(1, 2, 3, 4, NA, NA, NA, 5, 6, 7, NA),
+                  y = c(NA, NA, NA, 1, 2, 3, 4, 5, 6, 7, 8))
+    expect_equal(join(x, y, 3, 0, type = "outer"), tol_3)
+    tol_3 <- tol_3[2:1]
+    names(tol_3) <- c("x", "y")
+    expect_equal(join(y, x, 3, 0, type = "outer"), tol_3)
+
+    x <- c(1, 1.5, 2, 2.1, 3, 3.4, 4.1, 4.5, 5, 5.1, 5.2, 6, 14)
+    y <- c(4.6, 4.7, 4.8, 4.9, 5, 6, 7, 8)
+    tol_3 <- list(
+        x = c(1, 2, 3, 4, 5, 6, 7, 8, NA, NA, NA, 9, 10, 11, 12, NA, NA, 13),
+        y = c(NA, NA, NA, NA, NA, NA, NA, 1, 2, 3, 4, 5, NA, NA, 6, 7, 8, NA))
+    expect_equal(join(x, y, 3, 0, type = "outer"), tol_3)
+
+    x <- c(1, 2, 3, 4)
+    y <- c(6, 7, 8)
+    join(x, y, 20, 0, type = "outer")
+    expect_equal(join(x, y, 0, 0, type = "outer"),
+                 list(x = c(1:4, NA, NA, NA), y = c(NA, NA, NA, NA, 1:3)))
+    expect_equal(join(x, y, 20, 0, type = "outer"),
+                 list(x = c(1:4, NA, NA), y = c(NA, NA, NA, 1:3)))
+
+    ## no match at all
+    x <- as.numeric(c(1, 2, 3, 6))
+    y <- as.numeric(c(4, 5, 7))
+    expect_equal(join(x, y, 0, 0, type = "outer"),
+                 list(x = c(1:3, NA, NA, 4, NA), y = c(NA, NA, NA, 1:2, NA, 3)))
+
+    ## identical values
+    x <- c(3, 5, 5, 5, 6, 9)
+    y <- c(2, 4, 5, 5.5, 7)
+    expect_equal(join(x, y, 0, 0, type = "outer"),
+                 list(x = c(NA, 1, NA, 2, 3, 4, NA, 5, NA, 6),
+                      y = c(1, NA, 2, 3, NA, NA, 4, NA, 5, NA)))
+    expect_equal(join(y, x, 0, 0, type = "outer"),
+                 list(x = c(1, NA, 2, 3, NA, NA, 4, NA, 5, NA),
+                      y = c(NA, 1, NA, 2, 3, 4, NA, 5, NA, 6)))
+})
+
+test_that("left join works", {
+    x <- as.numeric(c(1, 2, 3, 6))
+    y <- as.numeric(c(3, 4, 5, 6, 7))
+
+    expect_error(join(1:3, 4:1, 0, 0, type = "left"), "sorted")
+    expect_error(join(4:1, 1:4, 0, 0, type = "left"), "sorted")
+    expect_error(join(c(1, 2, NA, 3, type = "left"), 1:4, 0, 0), "sorted")
+
+    expect_equal(join(x, y, tolerance = 0, ppm = 0, type = "left"),
+                 list(x = 1:4, y = c(NA, NA, 1, 4)))
+    expect_equal(join(x, y, tolerance = 5, ppm = 0, type = "left"),
+                 list(x = 1:4, y = c(NA, NA, 1, 4)))
+
+    x <- as.numeric(c(1, 3, 5, 6, 8))
+    y <- as.numeric(c(3, 4, 5, 7))
+    expect_equal(join(x, y, 0, 0, type = "left"),
+                 list(x = 1:5, y = c(NA, 1, 3, NA, NA)))
+    expect_equal(join(y, x, 0, 0, type = "left"),
+                 list(x = 1:4, y = c(2, NA, 3, NA)))
+    expect_equal(join(x, y, 1, 0, type = "left"),
+                 list(x = 1:5, y = c(NA, 1, 3, 4, NA)))
+    expect_equal(join(y, x, 1, 0, type = "left"),
+                 list(x = 1:4, y = c(2, NA, 3, 4)))
+
+    x <- c(133.0759, 133.0775, 133.9788, 133.9804, 133.9820, 133.9837)
+    y <- c(133.9755, 133.9771, 133.9788, 133.9804, 133.9820, 133.9836)
+    expect_equal(join(x, y, 0, 0, type = "left"),
+                 list(x = 1:6, y = c(NA, NA, 3, 4, 5, NA)))
+    expect_equal(join(x, y, 0.01, 0, type = "left"),
+                 list(x = 1:6, y = c(NA, NA, 3, 4, 5, 6)))
+    expect_equal(join(x, y, 1, 0, type = "left"),
+                 list(x = 1:6, y = c(NA, 1, 3, 4, 5, 6)))
+
+    x <- c(1, 1.5, 2, 2.1, 5, 6, 7)
+    y <- c(4.6, 4.7, 4.8, 4.9, 5, 6, 7, 8)
+    expect_equal(join(x, y, 0, 0, type = "left"),
+                 list(x = 1:7, y = c(NA, NA, NA, NA, 5, 6, 7)))
+    expect_equal(join(x, y, 1, 0, type = "left"),
+                 list(x = 1:7, y = c(NA, NA, NA, NA, 5, 6, 7)))
+    expect_equal(join(x, y, 3, 0, type = "left"),
+                 list(x = 1:7, y = c(NA, NA, NA, 1, 5, 6, 7)))
+    expect_equal(join(x, y, 10, 0, type = "left"),
+                 list(x = 1:7, y = c(NA, NA, NA, 1, 5, 6, 7)))
+    expect_equal(join(y, x, 0, 0, type = "left"),
+                 list(x = 1:8, y = c(NA, NA, NA, NA, 5, 6, 7, NA)))
+    expect_equal(join(y, x, 10, 0, type = "left"),
+                 list(x = 1:8, y = c(NA, NA, NA, NA, 5, 6, 7, NA)))
 })